Improve Windows installation instructions
This is based on an interaction with a user on the Discord, who was not sure if the output of the psp-config command was correct or how to open an Ubunutu terminal in a specific location. This change should address that.

When I tried to do the install on my Windows installation, I also found curl did not download the release, but wget did. So I changed to wget.