Fix cmake deprecation warning
CMake release 3.31 deprecates compatibility with versions of CMake older than 3.10. The release notes states: 'Compatibility with versions of CMake older than 3.10 is now deprecated and will be removed from a future version. Calls to cmake_minimum_required() or cmake_policy() that set the policy version to an older value now issue a deprecation diagnostic.'

https://cmake.org/cmake/help/latest/release/3.31.html#deprecated-and-removed-features

Updating from 3.5 release (2016-03-08) to 3.11 release (2018-03-28) fixes the warning without giving an high constraint to the users because it will work for install 7 years old.
Source code will not be blocked to old releases if other upgrades are needed in the future.